+++ lzero/entry/utils.py
@@ -49,27 +49,29 @@ def log_buffer_memory_usage(train_iter: int, buffer: "GameBuffer", writer: Summa
         - buffer (:obj:`GameBuffer`): The game buffer.
         - writer (:obj:`SummaryWriter`): The TensorBoard writer.
     """
-    writer.add_scalar('Buffer/num_of_all_collected_episodes', buffer.num_of_collected_episodes, train_iter)
-    writer.add_scalar('Buffer/num_of_game_segments', len(buffer.game_segment_buffer), train_iter)
-    writer.add_scalar('Buffer/num_of_transitions', len(buffer.game_segment_game_pos_look_up), train_iter)
+    # "writer is None" means we are in a slave process in the DDP setup.
+    if writer is not None:
+        writer.add_scalar('Buffer/num_of_all_collected_episodes', buffer.num_of_collected_episodes, train_iter)
+        writer.add_scalar('Buffer/num_of_game_segments', len(buffer.game_segment_buffer), train_iter)
+        writer.add_scalar('Buffer/num_of_transitions', len(buffer.game_segment_game_pos_look_up), train_iter)
 
-    game_segment_buffer = buffer.game_segment_buffer
+        game_segment_buffer = buffer.game_segment_buffer
 
-    # Calculate the amount of memory occupied by self.game_segment_buffer (in bytes).
-    buffer_memory_usage = asizeof(game_segment_buffer)
+        # Calculate the amount of memory occupied by self.game_segment_buffer (in bytes).
+        buffer_memory_usage = asizeof(game_segment_buffer)
 
-    # Convert buffer_memory_usage to megabytes (MB).
-    buffer_memory_usage_mb = buffer_memory_usage / (1024 * 1024)
+        # Convert buffer_memory_usage to megabytes (MB).
+        buffer_memory_usage_mb = buffer_memory_usage / (1024 * 1024)
 
-    # Record the memory usage of self.game_segment_buffer to TensorBoard.
-    writer.add_scalar('Buffer/memory_usage/game_segment_buffer', buffer_memory_usage_mb, train_iter)
+        # Record the memory usage of self.game_segment_buffer to TensorBoard.
+        writer.add_scalar('Buffer/memory_usage/game_segment_buffer', buffer_memory_usage_mb, train_iter)
 
-    # Get the amount of memory currently used by the process (in bytes).
-    process = psutil.Process(os.getpid())
-    process_memory_usage = process.memory_info().rss
+        # Get the amount of memory currently used by the process (in bytes).
+        process = psutil.Process(os.getpid())
+        process_memory_usage = process.memory_info().rss
 
-    # Convert process_memory_usage to megabytes (MB).
-    process_memory_usage_mb = process_memory_usage / (1024 * 1024)
+        # Convert process_memory_usage to megabytes (MB).
+        process_memory_usage_mb = process_memory_usage / (1024 * 1024)
 
-    # Record the memory usage of the process to TensorBoard.
-    writer.add_scalar('Buffer/memory_usage/process', process_memory_usage_mb, train_iter)
+        # Record the memory usage of the process to TensorBoard.
+        writer.add_scalar('Buffer/memory_usage/process', process_memory_usage_mb, train_iter)