+++ zoo/board_games/tictactoe/config/tictactoe_alphazero_bot_mode_config.py
@@ -42,7 +42,7 @@ tictactoe_alphazero_config = dict(
         mcts_ctree=mcts_ctree,
         # ==============================================================
         # for the creation of simulation env
-        simulation_env_name='tictactoe',
+        simulation_env_id='tictactoe',
         simulation_env_config_type='play_with_bot',
         # ==============================================================
         model=dict(