+++ zoo/game_2048/entry/2048_bot_eval.py
@@ -7,7 +7,7 @@ from zoo.game_2048.envs.game_2048_env import Game2048Env
 
 # Define game configuration
 config = EasyDict(dict(
-    env_name="game_2048",
+    env_id="game_2048",
     # (str) The render mode. Options are 'None', 'state_realtime_mode', 'image_realtime_mode' or 'image_savefile_mode'.
     # If None, then the game will not be rendered.
     render_mode='image_realtime_mode',